+++ cmd/parser/gas/gulf_parser.go
@@ -62,7 +62,7 @@ func (g GasGulf) GetData() ([]Item, error) {
 		results = append(results, Item{
 			Name:  name,
 			Tag:   tag,
-			Price: price,
+			Price: int64(price),
 			Date:  time.Now().Format("2006-01-02"),
 		})
 	})
