+++ cmd/parser/gas/lukoili_parser.go
@@ -86,7 +86,7 @@ func (g GasLukoili) GetData() ([]Item, error) {
 		results = append(results, Item{
 			Name:  tag.Name,
 			Tag:   tag.Tag,
-			Price: price,
+			Price: int64(price),
 			Date:  time.Now().Format("2006-01-02"),
 		})
 	})
