+++ cmd/parser/gas/rompetrol_parser.go
@@ -60,7 +60,7 @@ func (g GasRompetrol) GetData() ([]Item, error) {
 		results = append(results, Item{
 			Name:  name,
 			Tag:   tag,
-			Price: price,
+			Price: int64(price),
 			Date:  time.Now().Format("2006-01-02"),
 		})
 	})
